Move error handler tests into main test file
